fix: fix data loss issue
