+++ src/events/index.ts
@@ -1,2 +1,2 @@
 export * from './user.event';
-export * from './trip.event'
+export * from './trip.event';
