+++ src/trips/trip.module.ts
@@ -3,7 +3,6 @@ import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { LocationRepository } from 'locations/location.repository';
-import { AMQPHandler } from '../rabbitmq';
 import { UserRepository } from 'users/user.repository';
 import { JwtStrategy } from './strategy';
 import { TripController } from './trip.controller';
@@ -33,6 +32,6 @@ import { config } from '../../config';
     TypeOrmModule.forFeature([TripRepository, LocationRepository, UserRepository]),
   ],
   controllers: [TripController],
-  providers: [TripService, JwtStrategy, AMQPHandler, TripEventSubscribers],
+  providers: [TripService, JwtStrategy, TripEventSubscribers],
 })
 export class TripModule {}
